Rename dependency_graph to config.
It is for configuration, not for dependency-graphing!
