+++ bin/xmra.py
@@ -10,7 +10,7 @@ import os
 import datetime
 from xmra.repositories.local.model import Library
 from xmra.xonotic.objects import MapPackage
-from xmra.dependency_graph import config
+from xmra.config import config
 from xmra.repositories.local.db import session
 from xmra.helpers.database import get_or_create
 from xmra.util import DateTimeEncoder
